pc - update github actions that use JamesIves/github-pages-deploy-action@v4 to try 3 times before giving up
The action JamesIves/github-pages-deploy-action@v4 often fails because of contention for the gh-pages branch.
In this PR, we wrap many of these calls in a call to the Wandalen/wretry.action@master action,

which will try multiple times before giving up.

We hope this results in far fewer CI/CD jobs failing.